Allow several threads to safely share a contact point if history is disabled
Previously, it was difficult to use the openCRG library in a thread-safe
manner, since creation of a contact point modifies a global data structure
that is read e.g. by crgEvaluv2z. That means that it is impossible to
create a new contact point in a thread while another thread queries the
OpenCRG data. But sharing a contact point was also impossible, since the
query writes into the contact point.

This patch changes the library to not modify the contact point if the
history is disabled (which can be done e.g. via the new function
crgContactPointSetHistoryForAllPoints). Thus contact points with disabled
history can be safely shared across threads.

Switch to C99.